Move all compilation unit-related methods into a separate helper class
